+++ packages/shared_prefs_typed/test/src/error_no_underscore.dart
@@ -1,0 +1,8 @@
+// This file is used as test input for error case validation.
+
+import 'package:shared_prefs_typed_annotations/shared_prefs_typed_annotations.dart';
+
+@TypedPrefs()
+abstract class AppPreferences {
+  static const int counter = 0;
+}