+++ packages/shared_prefs_typed/test/src/interface_case.dart
@@ -1,0 +1,9 @@
+// ignore_for_file: unused_field, unused_element // Specification for code generation only.
+
+import 'package:shared_prefs_typed_annotations/shared_prefs_typed_annotations.dart';
+
+@TypedPrefs(generateInterface: true)
+abstract class _InterfacePrefs {
+  static const int counter = 0;
+  static const String? name = null;
+}